+++ examples/demo_bpy_1a/client_app.py
@@ -1,0 +1,2 @@
+from client.app import * # noqa
+# import client.app