+++ gladius/starter/__init__.py
@@ -1,0 +1,1 @@
+from .old_starter import * # noqa